+++ packages/app-bridge-app/CHANGELOG.md
@@ -1,1 +1,9 @@
 # @frontify/app-bridge-app
+
+## 0.0.2
+
+### Patch Changes
+
+-   [#986](https://github.com/Frontify/brand-sdk/pull/986) [`6adb54b`](https://github.com/Frontify/brand-sdk/commit/6adb54bb2a3af14a016b3d979f1c43129953b5f4) Thanks [@julianiff](https://github.com/julianiff)! - feat: copy appBridgePlatformApp to app-bridge-app
+
+-   [#990](https://github.com/Frontify/brand-sdk/pull/990) [`e1ec08a`](https://github.com/Frontify/brand-sdk/commit/e1ec08a988950997ddc70213300639fd8f3866d9) Thanks [@julianiff](https://github.com/julianiff)! - fix: fix linting and tests